Separates caching implementation into separate files, replacing tests with common cache store tests (and separate cache ctrl tests, todo), and updates ram cache store with clear method and other recent changes. S3 cache store needs testing.
